// src/features/invoice/types/deal-types.ts
export type Deal = {
  acv: string
  businessMonth: string
  color: string
  name: string
  dealDate: string
  dealNo: string
  dealType: string
  totalGross: string
  interiorColor: string
  make: string
  miles: string
  model: string
  modelNo: string
  saleType: string
  relayDealType: string
  salesperson1: string
  salesperson2: string
  fiMgr1: string
  dealSource: string
  stockNo: string
  vehicleStatus: string
  verificationId: string
  year: string
  contractDate: string
  vehicleMileage: string
  totalTradesACV: string
  homePhone: string
  customerPhone?:string
}